backend/python-service/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pdfminer.six==20231228
spacy==3.8.4
python-dotenv==1.0.1
python-multipart==0.0.20
